Restructure 🧱
